fix: participation 관련 패키지 변경
