+++ src/main/java/com/gagoo/thiscoding/domain/maria/coderoom/controller/ParticipationController.java
@@ -1,8 +1,8 @@
-package com.gagoo.thiscoding.domain.maria.usercoderoom.controller;
+package com.gagoo.thiscoding.domain.maria.coderoom.controller;
 
+import com.gagoo.thiscoding.domain.maria.coderoom.controller.port.ParticipationService;
+import com.gagoo.thiscoding.domain.maria.coderoom.controller.response.ParticipatingCodeRoomResponse;
 import com.gagoo.thiscoding.domain.maria.user.domain.contants.Role;
-import com.gagoo.thiscoding.domain.maria.usercoderoom.controller.port.UserCodeRoomService;
-import com.gagoo.thiscoding.domain.maria.usercoderoom.controller.response.ParticipationsResponse;
 import com.gagoo.thiscoding.global.paging.dto.PageResponse;
 import com.gagoo.thiscoding.global.security.aop.AuthorizationRequired;
 import lombok.RequiredArgsConstructor;
@@ -13,16 +13,16 @@ import org.springframework.web.bind.annotation.*;
 import static org.springframework.http.HttpStatus.OK;
 
 @RestController
-@RequestMapping("/api/participations")
+@RequestMapping("/api/my-page/participations")
 @RequiredArgsConstructor
-public class UserCodeRoomParticipationsController {
-    private final UserCodeRoomService userCodeRoomService;
+public class ParticipationController {
+    private final ParticipationService participationService;
 
     // 참여 중인 코드방 목록 조회
     @GetMapping
     @AuthorizationRequired(value = Role.USER, status = OK)
-    public ResponseEntity<PageResponse<ParticipationsResponse>> getParticipationList(@RequestParam(defaultValue = "1") int page) {
-        Page<ParticipationsResponse> userCodeRooms = userCodeRoomService.getParticipations(page);
+    public ResponseEntity<PageResponse<ParticipatingCodeRoomResponse>> getParticipationList(@RequestParam(defaultValue = "1") int page) {
+        Page<ParticipatingCodeRoomResponse> userCodeRooms = participationService.getParticipations(page);
         return ResponseEntity
                 .ok()
                 .body(PageResponse.create(userCodeRooms));
@@ -32,7 +32,7 @@ public class UserCodeRoomParticipationsController {
     @PatchMapping("/{id}/access")
     @AuthorizationRequired(value = Role.USER, status = OK)
     public ResponseEntity<?> accessParticipation(@PathVariable Long id) {
-        userCodeRoomService.accessUserCodeRoom(id);
+        participationService.accessUserCodeRoom(id);
         return ResponseEntity
                 .ok()
                 .body("코드방 입/퇴장 성공");
@@ -42,7 +42,7 @@ public class UserCodeRoomParticipationsController {
     @DeleteMapping("/{id}/leave")
     @AuthorizationRequired(value = Role.USER, status = OK)
     public ResponseEntity<?> leaveParticipation(@PathVariable Long id) {
-        userCodeRoomService.leaveUserCodeRoom(id);
+        participationService.leaveUserCodeRoom(id);
         return ResponseEntity
                 .ok()
                 .body("코드방 탈퇴 성공");
