+++ src/main/java/com/gagoo/thiscoding/domain/maria/coderoom/controller/response/ParticipatingCodeRoomResponse.java
@@ -1,4 +1,4 @@
-package com.gagoo.thiscoding.domain.maria.usercoderoom.controller.response;
+package com.gagoo.thiscoding.domain.maria.coderoom.controller.response;
 
 import com.gagoo.thiscoding.domain.maria.coderoom.domain.CodeRoom;
 import com.gagoo.thiscoding.domain.maria.user.domain.User;
@@ -6,12 +6,11 @@ import com.gagoo.thiscoding.domain.maria.usercoderoom.domain.UserCodeRoom;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Getter
 @Builder
-public class ParticipationsResponse {
+public class ParticipatingCodeRoomResponse {
     private Long userCodeRoomId;
     private Long codeRoomId;
     private String title;
@@ -21,10 +20,10 @@ public class ParticipationsResponse {
     private int headCount;
 
     @Builder
-    public static ParticipationsResponse from(UserCodeRoom userCodeRoom, List<User> userList) {
+    public static ParticipatingCodeRoomResponse from(UserCodeRoom userCodeRoom, List<User> userList) {
         CodeRoom codeRoom = userCodeRoom.getCodeRoom();
 
-        return ParticipationsResponse.builder()
+        return ParticipatingCodeRoomResponse.builder()
                 .userCodeRoomId(userCodeRoom.getId())
                 .codeRoomId(codeRoom.getId())
                 .title(codeRoom.getTitle())
