+++ src/main/java/com/gagoo/thiscoding/domain/maria/usercoderoom/service/port/UserCodeRoomRepository.java
@@ -4,6 +4,8 @@ import com.gagoo.thiscoding.domain.maria.usercoderoom.domain.UserCodeRoom;
 import java.util.Optional;
 
 public interface UserCodeRoomRepository {
+    UserCodeRoom getById(Long codeRoomId);
+
     Optional<UserCodeRoom> findById(Long codeRoomId);
 
     UserCodeRoom save(UserCodeRoom userCodeRoom);
